requirements.txt
absl-py==0.2.1
asn1crypto==0.24.0
astor==0.6.2
bleach==1.5.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
Cython==0.28.4
enum34==1.1.6
Flask==1.0.2
future==0.16.0
gast==0.2.0
grpcio==1.11.0
gunicorn==19.9.0
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
joblib==0.12.0
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.0
pandas==0.23.3
pathlib==1.0.1
plac==0.9.6
protobuf==3.6.1
pycparser==2.18
python-crfsuite==0.9.5
python-dateutil==2.7.3
pytz==2018.5
scikit-learn==0.19.1
scikit-surprise==1.0.6
scipy==1.0.0
semantic-version==2.6.0
six==1.11.0
sklearn-crfsuite==0.3.6
snips-nlu==0.19.2
snips-nlu-ontology==0.62.0
snips-nlu-utils==0.8.0
surprise==0.1
tabulate==0.8.2
termcolor==1.1.0
tqdm==4.23.4
Werkzeug==0.14.1
wincertstore==0.2
flask-cors
flask-restful
pydub
google-cloud-texttospeech